Create a dataholder object
