+++ apps/student/src/main/java/com/instructure/student/activity/CallbackActivity.kt
@@ -21,6 +21,8 @@ import android.os.Bundle
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.heapanalytics.android.Heap
 import com.instructure.canvasapi2.StatusCallback
+import com.instructure.canvasapi2.apis.CourseAPI
+import com.instructure.canvasapi2.builders.RestParams
 import com.instructure.canvasapi2.managers.FeaturesManager
 import com.instructure.canvasapi2.managers.LaunchDefinitionsManager
 import com.instructure.canvasapi2.managers.ThemeManager
@@ -42,6 +44,7 @@ import com.instructure.canvasapi2.utils.ApiType
 import com.instructure.canvasapi2.utils.LinkHeaders
 import com.instructure.canvasapi2.utils.LocaleUtils
 import com.instructure.canvasapi2.utils.Logger
+import com.instructure.canvasapi2.utils.depaginate
 import com.instructure.canvasapi2.utils.pageview.PandataInfo
 import com.instructure.canvasapi2.utils.pageview.PandataManager
 import com.instructure.canvasapi2.utils.weave.StatusCallbackError
@@ -59,6 +62,7 @@ import com.instructure.pandautils.utils.toast
 import com.instructure.student.BuildConfig
 import com.instructure.student.R
 import com.instructure.student.fragment.NotificationListFragment
+import com.instructure.student.router.EnabledTabs
 import com.instructure.student.service.StudentPageViewService
 import com.instructure.student.util.StudentPrefs
 import dagger.hilt.android.AndroidEntryPoint
@@ -72,6 +76,9 @@ abstract class CallbackActivity : ParentActivity(), OnUnreadCountInvalidated, No
 
     @Inject
     lateinit var featureFlagProvider: FeatureFlagProvider
+
+    @Inject
+    lateinit var courseApi: CourseAPI.CoursesInterface
 
     private var loadInitialDataJob: Job? = null
 
@@ -89,6 +96,11 @@ abstract class CallbackActivity : ParentActivity(), OnUnreadCountInvalidated, No
     private fun loadInitialData() {
         loadInitialDataJob = tryWeave {
             setupHeapTracking()
+
+            // Get enabled tabs
+            EnabledTabs.enabledTabs = courseApi.getFirstPageCourses(RestParams()).depaginate {
+                courseApi.next(it, RestParams())
+            }.dataOrNull?.mapNotNull { it.tabs }?.flatten() ?: emptyList()
 
             // Determine if user can masquerade
             if (ApiPrefs.canBecomeUser == null) {
